story(avroc): drop explicit Windows support
avroc is distributed as a scratch base image and a Dagger module, neither of
which puts a Windows host in the position of running a generator executable.
The cross-platform machinery behind generator discovery was maintained surface
with no consumer behind it, so it goes:

- avroc_windows.go and its test are deleted.
- avroc_unix.go folds into avroc.go as ordinary code; isGeneratorExecutable
  now sits beside its only caller, and its test folds into avroc_test.go as a
  subtest of TestLookupGenerators.
- generatorKey is removed rather than kept as a no-op. It was the identity
  function once the .exe suffix was gone, so lookupGenerators keys the index
  on the entry name directly.
- docs/plugin/SPEC.md gains a Host platform section recording the decision, so
  a plugin author knows what they are writing against.
- The stale .exe entries in .gitignore and the Windows rationale on two test
  comments go with them.

Closes #104
