--- internal/avroc/avroc_windows.go
@@ -1,22 +1,0 @@
-// Copyright (c) 2026 Z5Labs and Contributors
-//
-// This software is released under the MIT License.
-// https://opensource.org/licenses/MIT
-
-//go:build windows
-
-package avroc
-
-import (
-	"io/fs"
-	"path/filepath"
-	"strings"
-)
-
-func isGeneratorExecutable(name string, mode fs.FileMode) bool {
-	return mode.IsRegular() && strings.HasPrefix(name, "avroc-gen-") && strings.EqualFold(filepath.Ext(name), ".exe")
-}
-
-func generatorKey(name string) string {
-	return strings.TrimSuffix(name, filepath.Ext(name))
-}
